fix: tiny change

+++ src/code/comb/ball_box_dda.hpp
@@ -6,7 +6,7 @@
 namespace tifa_libs::math {
 
 template <class mint>
-constexpr mint ball_box_dda(u64 ball, u64 box) { return qpow<mint>(box, ball); }
+constexpr mint ball_box_dda(u32 ball, u32 box) { return qpow<mint>(box, ball); }
 
 }  // namespace tifa_libs::math
