+++ src/code/comb/ball_box_dil.hpp
@@ -6,9 +6,9 @@
 namespace tifa_libs::math {
 
 template <class T>
-constexpr auto ball_box_dil(u64 ball, u64 box, poly<T> const& s2r) { return box > ball ? 0 : s2r[box]; }
+constexpr auto ball_box_dil(u32 ball, u32 box, poly<T> const& s2r) { return box > ball ? 0 : s2r[box]; }
 template <class T>
-constexpr auto ball_box_dil(u64 ball, u64 box) { return ball_box_dil<T>(ball, box, gen_stirling2_row<T>(ball)); }
+constexpr auto ball_box_dil(u32 ball, u32 box) { return ball_box_dil<T>(ball, box, gen_stirling2_row<T>(ball)); }
 
 }  // namespace tifa_libs::math
