+++ src/code/comb/ball_box_idl.hpp
@@ -6,7 +6,7 @@
 namespace tifa_libs::math {
 
 template <class mint>
-inline mint ball_box_idl(u64 ball, u64 box, Binom<mint> const& binom) { return binom.mCn(ball - 1, box - 1); }
+inline mint ball_box_idl(u32 ball, u32 box, Binom<mint> const& binom) { return binom.mCn(ball - 1, box - 1); }
 
 }  // namespace tifa_libs::math
