+++ src/code/comb/ball_box_idm.hpp
@@ -6,7 +6,7 @@
 namespace tifa_libs::math {
 
 template <class mint>
-inline mint ball_box_idm(u64 ball, u64 box, Binom<mint> const& binom) { return binom.mCn(box, ball); }
+inline mint ball_box_idm(u32 ball, u32 box, Binom<mint> const& binom) { return binom.mCn(box, ball); }
 
 }  // namespace tifa_libs::math
