+++ src/code/comb/ball_box_iia.hpp
@@ -6,9 +6,9 @@
 namespace tifa_libs::math {
 
 template <class T>
-constexpr auto ball_box_iia(u64 ball, u64, poly<T> const& bbiif) { return bbiif[ball]; }
+constexpr auto ball_box_iia(u32 ball, u32, poly<T> const& bbiif) { return bbiif[ball]; }
 template <class T>
-constexpr auto ball_box_iia(u64 ball, u64 box) { return ball_box_iia<T>(ball, box, gen_ball_box_ii<T>(box, ball)); }
+constexpr auto ball_box_iia(u32 ball, u32 box) { return ball_box_iia<T>(ball, box, gen_ball_box_ii<T>(box, ball)); }
 
 }  // namespace tifa_libs::math
